Moved phreak package tests to drools-test-coverage/test-compiler-integration
