+++ drools-test-coverage/test-compiler-integration/src/test/java/org/drools/mvel/integrationtests/phreak/PhreakNotNodeTest.java
@@ -13,7 +13,7 @@
  * limitations under the License.
 */
 
-package org.drools.mvel.compiler.phreak;
+package org.drools.mvel.integrationtests.phreak;
 
 import org.drools.core.RuleBaseConfiguration;
 import org.drools.core.common.InternalWorkingMemory;
@@ -33,8 +33,8 @@ import org.drools.core.reteoo.builder.BuildContext;
 import org.drools.mvel.MVELDialectRuntimeData;
 import org.junit.Test;
 
-import static org.drools.mvel.compiler.phreak.A.a;
-import static org.drools.mvel.compiler.phreak.B.b;
+import static org.drools.mvel.integrationtests.phreak.A.a;
+import static org.drools.mvel.integrationtests.phreak.B.b;
 
 public class PhreakNotNodeTest {
