+++ drools-test-coverage/test-compiler-integration/src/test/java/org/drools/mvel/integrationtests/phreak/ScenarioTest.java
@@ -13,7 +13,7 @@
  * limitations under the License.
 */
 
-package org.drools.mvel.compiler.phreak;
+package org.drools.mvel.integrationtests.phreak;
 
 import org.drools.core.RuleBaseConfiguration;
 import org.drools.core.common.InternalWorkingMemory;
@@ -32,8 +32,8 @@ import org.drools.core.reteoo.builder.BuildContext;
 import org.drools.mvel.MVELDialectRuntimeData;
 import org.junit.Test;
 
-import static org.drools.mvel.compiler.phreak.B.b;
-import static org.drools.mvel.compiler.phreak.Pair.t;
+import static org.drools.mvel.integrationtests.phreak.B.b;
+import static org.drools.mvel.integrationtests.phreak.Pair.t;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
